feat(keyboard manager): fix cr comments

+++ packages/core/src/lib/managers/keyboard.manager.ts
@@ -15,10 +15,13 @@ export class KeyboardManager {
   }
 
   onKeyUp(event: KeyboardEvent) {
-    const deleteKeyPressedAction = new DeleteKeyPressed(this.engine, event);
-    this.engine.startFiringAction(deleteKeyPressedAction);
-    this.engine.fireAction();
-    this.engine.stopFiringAction();
+    // only if delete key bindings provided, set delete action
+    if (this.engine.getDiagramModel().getKeyBindings()?.delete) {
+      const deleteKeyPressedAction = new DeleteKeyPressed(this.engine, event);
+      this.engine.startFiringAction(deleteKeyPressedAction);
+      this.engine.fireAction();
+      this.engine.stopFiringAction();
+    }
   }
 
   onCopy() {
